OWVizGraph: Base class for OWFreeVizGraph and OWRadVizGraph
