+++ Orange/widgets/visualize/owfreeviz.py
@@ -19,12 +19,9 @@ from Orange.widgets import widget, gui, settings, report
 from Orange.widgets.utils.annotated_data import (
     create_annotated_table, ANNOTATED_DATA_SIGNAL_NAME, create_groups_table
 )
-from Orange.widgets.visualize.owscatterplotgraph import (
-    OWScatterPlotBase, OWProjectionWidget
-)
-from Orange.widgets.visualize.utils.plotutils import (
-    AnchorItem, MouseEventDelegate, VizInteractiveViewBox
-)
+from Orange.widgets.visualize.owscatterplotgraph import OWProjectionWidget
+from Orange.widgets.visualize.utils.component import OWVizGraph
+from Orange.widgets.visualize.utils.plotutils import AnchorItem
 from Orange.widgets.widget import Input, Output
 
 
@@ -146,43 +143,13 @@ class AsyncUpdateLoop(QObject):
             super().customEvent(event)
 
 
-class OWFreeVizGraph(OWScatterPlotBase):
+class OWFreeVizGraph(OWVizGraph):
     radius = settings.Setting(0)
 
     def __init__(self, scatter_widget, parent):
-        super().__init__(scatter_widget, parent, VizInteractiveViewBox)
-        self._tooltip_delegate = MouseEventDelegate(self.help_event,
-                                                    self._show_indicator_event)
-        self.plot_widget.scene().installEventFilter(self._tooltip_delegate)
-
+        super().__init__(scatter_widget, parent)
         self._points = []
-        self._attributes = ()
-
         self._point_items = []
-        self._circle_item = None
-        self._indicator_item = None
-
-    def update_coordinates(self):
-        super().update_coordinates()
-        self._update_items()
-        self.view_box.setRange(RANGE)
-        self.view_box.setAspectLocked(True, 1)
-
-    def reset_button_clicked(self):
-        self.view_box.setRange(RANGE)
-
-    def set_attributes(self, attributes):
-        self._attributes = attributes
-
-    def set_point(self, i, x, y):
-        self._points[i][0] = x
-        self._points[i][1] = y
-
-    def set_points(self, points):
-        self._points = points
-
-    def get_points(self):
-        return self._points
 
     def update_radius(self):
         if self._circle_item is None:
@@ -193,8 +160,8 @@ class OWFreeVizGraph(OWScatterPlotBase):
             axitem.setVisible(np.linalg.norm(point) > r)
         self._circle_item.setRect(QRectF(-r, -r, 2 * r, 2 * r))
 
-    def show_indicator(self, point_i):
-        self._update_indicator_item(point_i)
+    def set_view_box_range(self):
+        self.view_box.setRange(RANGE)
 
     def can_show_indicator(self, pos):
         if not len(self._points):
@@ -205,51 +172,25 @@ class OWFreeVizGraph(OWScatterPlotBase):
         mask[np.linalg.norm(self._points, axis=1) > r] = True
         distances = distance.cdist([[pos.x(), pos.y()]], self._points)[0]
         distances = distances[mask]
-        if len(distances) and np.min(distances) < 0.08:
+        if len(distances) and np.min(distances) < self.DISTANCE_DIFF:
             return True, np.flatnonzero(mask)[np.argmin(distances)]
         return False, None
 
-    def _show_indicator_event(self, ev):
-        scene = self.plot_widget.scene()
-        if self.scatterplot_item is None or scene.drag_tooltip.isVisible():
-            return False
+    def _remove_point_items(self):
+        for item in self._point_items:
+            self.plot_widget.removeItem(item)
+        self._point_items = []
 
-        pos = self.scatterplot_item.mapFromScene(ev.scenePos())
-        can_show, point_i = self.can_show_indicator(pos)
-        if can_show:
-            self._update_indicator_item(point_i)
-            if self.view_box.mouse_state == 0:
-                self.view_box.setCursor(Qt.OpenHandCursor)
-        else:
-            if self._indicator_item is not None:
-                self.plot_widget.removeItem(self._indicator_item)
-                self._indicator_item = None
-            if self.view_box.mouse_state == 0:
-                self.view_box.setCursor(Qt.ArrowCursor)
-        return True
+    def _add_point_items(self):
+        r = self.radius / 100 + 1e-5
+        for point, var in zip(self._points, self._attributes):
+            axitem = AnchorItem(line=QLineF(0, 0, *point), text=var.name)
+            axitem.setVisible(np.linalg.norm(point) > r)
+            axitem.setPen(pg.mkPen((100, 100, 100)))
+            self.plot_widget.addItem(axitem)
+            self._point_items.append(axitem)
 
-    def _update_items(self):
-        self._update_point_items()
-        self._update_circle_item()
-
-    def _update_point_items(self):
-        self.__remove_point_items()
-        self.__add_point_items()
-
-    def _update_circle_item(self):
-        self.__remove_circle_item()
-        self.__add_circle_item()
-
-    def _update_indicator_item(self, point_i):
-        self.__remove_indicator_item()
-        self.__add_indicator_item(point_i)
-
-    def __remove_circle_item(self):
-        if self._circle_item is not None:
-            self.plot_widget.removeItem(self._circle_item)
-            self._circle_item = None
-
-    def __add_circle_item(self):
+    def _add_circle_item(self):
         if not len(self._points):
             return
         r = self.radius / 100 + 1e-5
@@ -259,26 +200,7 @@ class OWFreeVizGraph(OWScatterPlotBase):
         self._circle_item.setPen(pen)
         self.plot_widget.addItem(self._circle_item)
 
-    def __remove_point_items(self):
-        for axitem in self._point_items:
-            self.plot_widget.removeItem(axitem)
-        self._point_items = []
-
-    def __add_point_items(self):
-        r = self.radius / 100 + 1e-5
-        for point, var in zip(self._points, self._attributes):
-            axitem = AnchorItem(line=QLineF(0, 0, *point), text=var.name)
-            axitem.setVisible(np.linalg.norm(point) > r)
-            axitem.setPen(pg.mkPen((100, 100, 100)))
-            self.plot_widget.addItem(axitem)
-            self._point_items.append(axitem)
-
-    def __remove_indicator_item(self):
-        if self._indicator_item is not None:
-            self.plot_widget.removeItem(self._indicator_item)
-            self._indicator_item = None
-
-    def __add_indicator_item(self, point_i):
+    def _add_indicator_item(self, point_i):
         x, y = self._points[point_i]
         dx = (self.view_box.childGroup.mapToDevice(QPoint(1, 0)) -
               self.view_box.childGroup.mapToDevice(QPoint(-1, 0))).x()
